requirements.txt
opencv-python
open3d-python==0.7.0.0
scikit-learn
tqdm
shapely
